Sonus-Bot/requirements.txt
discord.py>=2.4.0
yt-dlp
aiohttp
python-dotenv
pydantic
asyncpg
requests
# Test dependencies
pytest
pytest-asyncio
